chore(FormattingToolbar): modularize popup calc code - I199
