+++ src/FormattingToolbar/index.js
@@ -10,6 +10,7 @@ import * as action from './toolbarMethods';
 import * as styles from './toolbarStyles';
 import * as tips from './toolbarTooltip';
 import * as CONST from '../constants';
+import calculateLinkPopupPosition from './LinkComponent';
 
 import * as boldIcon from '../icons/bold';
 import * as italicIcon from '../icons/italic';
@@ -101,7 +102,6 @@ export default class FormatToolbar extends React.Component {
     this.removeLinkForm = this.removeLinkForm.bind(this);
     this.closeSetLinkForm = this.closeSetLinkForm.bind(this);
     this.renderLinkSetForm = this.renderLinkSetForm.bind(this);
-    this.calculateLinkPopupPosition = this.calculateLinkPopupPosition.bind(this);
   }
 
   componentDidMount() {
@@ -352,80 +352,14 @@ export default class FormatToolbar extends React.Component {
   }
 
   /**
-   * Calculates the link popup position and styles, if any
-   */
-  calculateLinkPopupPosition() {
-    // Constant values 2 and 20 (px) has been
-    // manually observed through devtools and what looked best.
-
-    let top = null;
-    let left = null;
-    let popupPosition = 'bottom center';
-
-    // No need to calculate position of the popup is it is not even opened!
-    // Same for if the current selection is not a link
-    const isLinkBool = action.hasLinks(this.props.editor);
-
-    const isLinkPopupOpened = this.state.openSetLink;
-    if (!isLinkPopupOpened && !action.isOnlyLink(this.props.editor)) {
-      return {
-        popupPosition,
-        // Hide the popup by setting negative zIndex
-        popupStyle: { zIndex: -1 }
-      };
-    }
-
-    if (isLinkBool && action.isOnlyLink(this.props.editor)) {
-      const linkInlineSelection = this.props.editor.value.inlines
-        .find(inline => inline.type === 'link');
-      this.props.editor.moveToRangeOfNode(linkInlineSelection);
-    }
-
-    // Get selection node from slate
-    const selection = this.props.editor.findDOMRange(this.props.editor.value.selection);
-
-    popupPosition = 'bottom left';
-
-    const { body, documentElement } = document;
-    const pageWidth = Math.max(body.scrollWidth, body.offsetWidth,
-      documentElement.clientWidth, documentElement.scrollWidth, documentElement.offsetWidth);
-
-    // Find the selected text position in DOM to place the popup relative to it
-    const rect = selection.getBoundingClientRect();
-
-    // distance from top of the document + the height of the element + scroll offet ...
-    // ... -2px to account for semantic-ui popup caret position
-    const CARET_TOP_OFFSET = 2;
-    top = rect.top + rect.height + window.scrollY - CARET_TOP_OFFSET;
-
-    // distance from the left of the document and ...
-    // ... subtracting 20px to account for the semantic-ui popup caret position
-    const calcMiddleSelection = (selection.endOffset - selection.startOffset) * 2;
-    const CARET_LEFT_OFFSET = (20 - calcMiddleSelection);
-    left = rect.left - CARET_LEFT_OFFSET;
-
-    const popupRect = this.setLinkFormPopup.getBoundingClientRect();
-
-    // Check if there is enough space on right, otherwise flip the popup horizontally
-    // and adjust the popup position accordingly
-    const spaceOnRight = pageWidth - rect.left;
-    if (spaceOnRight < popupRect.width) {
-      popupPosition = 'bottom right';
-      left = rect.left - popupRect.width + CARET_LEFT_OFFSET;
-    }
-    return {
-      // Disable semantic ui popup placement by overriding `transform`
-      // and use our computed `top` and `left` values
-      popupStyle: { top, left, transform: 'none' },
-      popupPosition,
-    };
-  }
-
-  /**
    * Render form in popup to set the link.
    */
   renderLinkSetForm() {
-    const { popupPosition, popupStyle } = this.calculateLinkPopupPosition();
+    const { popupPosition, popupStyle } = calculateLinkPopupPosition(
+      this.props.editor,
+      this.openSetLink,
+      this.setLinkFormPopup
+    );
     const { value } = this.props.editor;
     const { document, selection } = value;
     const isLinkBool = action.hasLinks(this.props.editor);
